--- gomen/src/main/java/com/tuna/gomen/bookMarkandLike/entity/BookmarkEntity.java
@@ -1,13 +1,0 @@
-package com.tuna.gomen.bookMarkandLike.entity;
-
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class BookmarkEntity {
-    private Integer postId;
-    private Integer userId;
-}
-
