+++ gomen/src/main/java/com/tuna/gomen/bookMarkandLike/query/dto/BookmarkedBoardDto.java
@@ -1,0 +1,16 @@
+package com.tuna.gomen.bookMarkandLike.query.dto;
+
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class BookmarkedBoardDto {
+    private Integer postId;
+    private String title;
+    private String isBlinded;
+    private String isDeleted;
+    private String userNickname;
+
+}
