+++ gomen/src/main/java/com/tuna/gomen/bookMarkandLike/query/entity/BookmarkEntity.java
@@ -1,0 +1,13 @@
+package com.tuna.gomen.bookMarkandLike.query.entity;
+
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class BookmarkEntity {
+    private Integer postId;
+    private Integer userId;
+}
+
