+++ gomen/src/main/java/com/tuna/gomen/comment/command/repository/CommentRepository.java
@@ -1,0 +1,12 @@
+package com.tuna.gomen.comment.command.repository;
+
+import com.tuna.gomen.comment.command.entity.Comment;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+
+public interface CommentRepository extends JpaRepository<Comment, Integer > {
+
+    List<Comment> findByBoard_PostIdAndIsDeleted(Integer postId, String isDeleted);
+
+}
