+++ gomen/src/main/java/com/tuna/gomen/mapper/BookmarkMapper.java
@@ -1,6 +1,6 @@
 package com.tuna.gomen.mapper;
 
-import com.tuna.gomen.bookMarkandLike.dto.BookmarkedBoardDto;
+import com.tuna.gomen.bookMarkandLike.query.dto.BookmarkedBoardDto;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Param;
 import java.util.List;
