+++ gomen/src/main/java/com/tuna/gomen/user/command/entity/User.java
@@ -1,6 +1,5 @@
 package com.tuna.gomen.user.command.entity;
 
-import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
