+++ gomen/src/main/java/com/tuna/gomen/user/command/service/UserService2.java
@@ -3,7 +3,9 @@ package com.tuna.gomen.user.command.service;
 import com.tuna.gomen.user.command.entity.User;
 import com.tuna.gomen.user.command.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
 public class UserService2 {
@@ -14,5 +16,7 @@ public class UserService2 {
         return userRepository.findById(userId)
                 .orElseThrow(() -> new RuntimeException("사용자를 찾을 수 없습니다. ID: " + userId));
     }
+
+
 
 }
